toolkit/components/places/tests/mochitest/test_bug_94514.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=94514
Specifically, this tests that a page that is obtained via a post request does
not get added to global history.
-->
<head>
  <title>Test for Bug 94515</title>
  <script type="text/javascript" src="http://mochi.test:8888/tests/SimpleTest/SimpleTest.js"></script>

  <link rel="stylesheet" type="text/css" href="http://mochi.test:8888/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=94514">Mozilla Bug 94514</a>

<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var startURI = "http://mochi.test:8888/tests/toolkit/components/places/tests/bug94514-postpage.html";
var postedURI = startURI + "?posted=1";

const Cc = Components.classes;
const Ci = Components.interfaces;

Components.utils.import("resource://gre/modules/PlacesUtils.jsm");

var ios = Cc["@mozilla.org/network/io-service;1"].
          getService(Ci.nsIIOService);
var startPage = ios.newURI(startURI, null, null);
var postedPage = ios.newURI(postedURI, null, null);
var w = null;

// Because adding visits is async, we will not be notified imemdiately.
var os = Cc["@mozilla.org/observer-service;1"].
         getService(Ci.nsIObserverService);
var visitObserver = {
  _visitCount: 0,
  observe: function(aSubject, aTopic, aData) {
    if (!startPage.equals(aSubject.QueryInterface(Ci.nsIURI)) ||
        ++this._visitCount < 2) {
      return;
    }
    os.removeObserver(this, aTopic);
    finishTest();
  },
};
os.addObserver(visitObserver, "uri-visit-saved", false);

PlacesUtils.asyncHistory.isURIVisited(startPage, function(aURI, aIsVisited) {
  SimpleTest.ok(!aIsVisited, "Initial page does not start in global history. " +
                "Note: this will also fail if you run the test twice.");
  PlacesUtils.asyncHistory.isURIVisited(postedPage, function(aURI, aIsVisited) {
    SimpleTest.ok(!aIsVisited, "Posted page does not start in global history.");
    w = window.open(startURI, "", "width=10,height=10");
  });
});

function finishTest() {
  // We need to check that this was not added to global history.
  PlacesUtils.asyncHistory.isURIVisited(startPage, function(aURI, aIsVisited) {
    SimpleTest.ok(aIsVisited, "Initial page was added to global history.");
    PlacesUtils.asyncHistory.isURIVisited(postedPage, function(aURI, aIsVisited) {
      SimpleTest.ok(!aIsVisited, "Posted page was not added to global history.");
      w.close();
      SimpleTest.finish();
    });
  });
}

</script>
</pre>
</body>
</html>
